adding cluster/standalone functions
