+++ python/python/glide/async_commands/standalone_commands.py
@@ -5,7 +5,12 @@ from __future__ import annotations
 from typing import Dict, List, Mapping, Optional, cast
 
 from glide.async_commands.command_args import Limit, OrderBy
-from glide.async_commands.core import CoreCommands, InfoSection, _build_sort_args
+from glide.async_commands.core import (
+    CoreCommands,
+    FlushMode,
+    InfoSection,
+    _build_sort_args,
+)
 from glide.async_commands.transaction import BaseTransaction, Transaction
 from glide.constants import TOK, TResult
 from glide.protobuf.redis_request_pb2 import RequestType
@@ -429,3 +434,28 @@ class StandaloneCommands(CoreCommands):
         """
         result = await self._execute_command(RequestType.Publish, [channel, message])
         return cast(int, result)
+
+    async def flushall(self, flush_mode: Optional[FlushMode] = None) -> TOK:
+        """
+        Deletes all the keys of all the existing databases. This command never fails.
+
+        See https://valkey.io/commands/flushall for more details.
+
+        Args:
+            flush_mode (Optional[FlushMode]): The flushing mode, could be either `SYNC` or `ASYNC`.
+
+        Returns:
+            str: OK.
+
+        Examples:
+             >>> await client.flushall(FlushMode.ASYNC)
+                 OK  # This command never fails.
+        """
+        args = []
+        if flush_mode is not None:
+            args.append(flush_mode.value)
+
+        return cast(
+            TOK,
+            await self._execute_command(RequestType.FlushAll, args),
+        )
