+++ python/python/tests/test_async_client.py
@@ -5151,13 +5151,21 @@ class TestCommands:
         key = f"{{key}}-1{get_random_string(5)}"
         value = get_random_string(5)
 
-        assert await redis_client.set(key, value)
+        await redis_client.set(key, value)
         assert await redis_client.dbsize() > 0
         assert await redis_client.flushall() is OK
         assert await redis_client.flushall(FlushMode.ASYNC) is OK
         if not await check_if_server_version_lt(redis_client, min_version):
             assert await redis_client.flushall(FlushMode.SYNC) is OK
         assert await redis_client.dbsize() == 0
+
+        if isinstance(redis_client, RedisClusterClient):
+            await redis_client.set(key, value)
+            assert await redis_client.flushall(route=AllPrimaries()) is OK
+            assert await redis_client.flushall(FlushMode.ASYNC, AllPrimaries()) is OK
+            if not await check_if_server_version_lt(redis_client, min_version):
+                assert await redis_client.flushall(FlushMode.SYNC, AllPrimaries()) is OK
+            assert await redis_client.dbsize() == 0
 
 
 class TestMultiKeyCommandCrossSlot:
